# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"DataResponse_string_": {
				"properties": {
					"data": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"ErrorResponse": {
				"properties": {
					"errors": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"errors",
					"status"
				],
				"type": "object"
			},
			"UserCreateRequest": {
				"properties": {
					"phone": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"phone",
					"password",
					"email",
					"name"
				],
				"type": "object"
			},
			"UserModel": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"uid": {
						"type": "string"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"phone",
					"email",
					"uid"
				],
				"type": "object"
			},
			"DataResponse_UserModel-Array_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserModel"
						},
						"type": "array"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"DataResponse_UserModel_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/UserModel"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"MessageResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"status"
				],
				"type": "object"
			},
			"UserUpdateRequest": {
				"properties": {
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object"
			}
		},
		"securitySchemes": {
			"X-API-TOKEN": {
				"type": "apiKey",
				"name": "X-API-TOKEN",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "Docs API Template",
		"version": "v1.0.0",
		"termsOfService": "https://example.com/tos",
		"contact": {
			"name": "Sendi Aagustian",
			"email": "sendiagustian.pk@gmail.com"
		},
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		}
	},
	"paths": {
		"/health-check": {
			"get": {
				"operationId": "Ping",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Health Check"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/user/create": {
			"post": {
				"operationId": "createUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/DataResponse_string_"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreateRequest"
							}
						}
					}
				}
			}
		},
		"/api/v1/user/all": {
			"get": {
				"operationId": "getAllUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/DataResponse_UserModel-Array_"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/user/{uid}": {
			"get": {
				"operationId": "getUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/DataResponse_UserModel_"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "uid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/user/update-data/{uid}": {
			"put": {
				"operationId": "updateDataUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/MessageResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "uid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/api/v1/user/delete/{uid}": {
			"delete": {
				"operationId": "deleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/MessageResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "uid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://127.0.0.1:5050",
			"description": "Local environment"
		},
		{
			"url": "http://127.0.0.1:8000",
			"description": "Prod environment"
		}
	]
}